requirements-test.txt
# pytest and testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
